created basic item, gem and enchant selection dropdowns
