+++ css/settings.css
@@ -218,7 +218,7 @@ ul {
   background-color: rgba(0,0,0,0.2); 
 
 }
-/* settings Modal Content */
+/* log Modal Content */
 .modal-log-content {
   position:absolute;
   left: 35%;
@@ -228,6 +228,31 @@ ul {
   border: 4px solid #2e2e33;
   width: 65%;
   height: auto;
+}
+.modal-gear {
+  display: none; 
+  position: fixed; 
+  z-index: 1; 
+  padding-top: 100px; 
+  top: 0;
+  width: 100%; 
+  height: 100%; 
+  overflow-y: auto; 
+  background-color: rgb(0,0,0); 
+  background-color: rgba(0,0,0,0.2); 
+
+}
+/* gear Modal Content */
+.modal-gear-content {
+  position:absolute;
+  left: 35%;
+  background-color: #1e1e22;
+  top: 40%;
+  margin: auto;
+  padding: 20px;
+  border: 4px solid #2e2e33;
+  width: 600px;
+  height: 300px;
 }
 
 /* The Close Button */
@@ -380,6 +405,11 @@ ul {
   font-size: 18px;
   height: 30px;
   width: 435px;
+}
+.gear-select-input {
+  margin-top: 2px;
+  margin-bottom: 2px;
+  width: 300px;
 }
 .fight-setting-input-1 {
   margin-top: 2px;
