+++ nicepage/nicepage.js
@@ -29,7 +29,7 @@ https://github.com/imakewebthings/waypoints/blob/master/licenses.txt
  * http://github.com/cferdinandi/gumshoe
  */
 !function(i,factory){if(true)n=[],o=function(){return factory(i)}.apply(e,n),!(void 0!==o&&(t.exports=o));else if("object"==typeof e)t.exports=factory(i);else i.Gumshoe=factory(i)}(void 0!==i?i:"undefined"!=typeof window?window:this,function(t){var e={navClass:"active",contentClass:"active",nested:false,nestedClass:"active",offset:0,reflow:false,events:true},i=function(){var t={};return Array.prototype.forEach.call(arguments,function(e){for(var i in e)if(e.hasOwnProperty(i))t[i]=e[i]}),t},n=function(type,t,e){if(e.settings.events){var i=new CustomEvent(type,{bubbles:true,cancelable:true,detail:e});t.dispatchEvent(i)}},o=function(t){var e=0;if(t.offsetParent)for(;t;)e+=t.offsetTop,t=t.offsetParent;return e>=0?e:0},a=function(t){if(t)t.sort(function(t,e){if(o(t.content)<o(e.content))return-1;else return 1})},s=function(settings){if("function"==typeof settings.offset)return parseFloat(settings.offset());else return parseFloat(settings.offset)},l=function(){return Math.max(document.body.scrollHeight,document.documentElement.scrollHeight,document.body.offsetHeight,document.documentElement.offsetHeight,document.body.clientHeight,document.documentElement.clientHeight)},u=function(e,settings,i){var n=e.getBoundingClientRect(),o=s(settings);if(i)return parseInt(n.bottom,10)<(t.innerHeight||document.documentElement.clientHeight);else return parseInt(n.top,10)<=o},f=function(){if(t.innerHeight+t.pageYOffset>=l())return true;else return false},c=function(t,settings){if(f()&&u(t.content,settings,true))return true;else return false},p=function(t,settings){if(t.length){var e=t[t.length-1];if(c(e,settings))return e;for(var i=t.length-1;i>=0;i--)if(u(t[i].content,settings))return t[i]}},h=function(nav,settings){if(settings.nested&&nav.parentNode){var t=nav.parentNode.closest("li");if(t)t.classList.remove(settings.nestedClass),h(t,settings)}},m=function(items,settings){if(items){var t=items.nav.closest("li");if(t)t.classList.remove(settings.navClass),items.content.classList.remove(settings.contentClass),h(t,settings),n("gumshoeDeactivate",t,{link:items.nav,content:items.content,settings:settings})}},v=function(nav,settings){if(settings.nested){var t=nav.parentNode.closest("li");if(t)t.classList.add(settings.nestedClass),v(t,settings)}},g=function(items,settings){if(items){var t=items.nav.closest("li");if(t)t.classList.add(settings.navClass),items.content.classList.add(settings.contentClass),v(t,settings),n("gumshoeActivate",t,{link:items.nav,content:items.content,settings:settings})}};return function(selector,n){var o={},s,l,u,f,settings;o.setup=function(){s=document.querySelectorAll(selector),l=[],Array.prototype.forEach.call(s,function(t){var e=document.getElementById(decodeURIComponent(t.hash.substr(1)));if(e)l.push({nav:t,content:e})}),a(l)},o.detect=function(){var t=p(l,settings);if(t){if(!u||t.content!==u.content)m(u,settings),g(t,settings),u=t}else if(u)m(u,settings),u=null};var c=function(){if(f)t.cancelAnimationFrame(f);f=t.requestAnimationFrame(o.detect)},h=function(){if(f)t.cancelAnimationFrame(f);f=t.requestAnimationFrame(function(){a(l),o.detect()})};return o.destroy=function(){if(u)m(u,settings);if(t.removeEventListener("scroll",c,false),settings.reflow)t.removeEventListener("resize",h,false);l=null,s=null,u=null,f=null,settings=null},function(){if(settings=i(e,n||{}),o.setup(),o.detect(),t.addEventListener("scroll",c,false),settings.reflow)t.addEventListener("resize",h,false)}(),o}})}).call(e,i(52))},6203:function(t,e,i){"use strict";var n=i(6),o=i(6204),HorizontalLayoutSlider=i(145);n(window).on("load",function(){setTimeout(function(){n(".u-gallery").removeClass("u-no-transition"),n(".u-layout-horizontal").each(function(){var gallery=n(this),slider=new HorizontalLayoutSlider(gallery,true);gallery.children(".u-gallery-nav").click(function(t){t.preventDefault();var e=n(t.currentTarget);slider.navigate(e)})})},250)}),n(function(){n("body").on("mouseenter",".u-gallery.u-no-transition",function(){n(this).closest(".u-gallery").removeClass("u-no-transition")}),new o([".u-gallery.u-product-zoom.u-layout-thumbnails",".u-gallery.u-product-zoom.u-layout-carousel"]).init()})},6204:function(t,e,i){"use strict";function n(t){this.galleryZoomSelector=t}function o(t){var e=t.currentTarget,i=l(e).closest(".u-gallery-item"),n=i.data("zoom_click"),o=e.getBoundingClientRect(),a=e.querySelector("img"),s=t.clientX,u=t.clientY,f=t.originalEvent.changedTouches;if(!n&&!f){l(e).addClass("hover");var c=s-o.x,p=u-o.y;requestAnimationFrame(function(){var t=c*(1-a.offsetWidth/e.offsetWidth),i=p*(1-a.offsetHeight/e.offsetHeight);a.style.left=t+"px",a.style.top=i+"px"})}}function a(t){var e=l(t.currentTarget);l(e).removeClass("hover"),l(e).closest(".u-gallery-item").data("zoom_click")}function s(t){var e=l(t.currentTarget);l(e).removeClass("hover")}var l=i(6);t.exports=n,n.prototype.init=function(){var t=this.galleryZoomSelector.map(function(selector){return selector+" .u-back-slide"}).join(", "),e=this.galleryZoomSelector.map(function(selector){return selector+" .u-back-image"}).join(", ");l("body").on("mousedown touchstart",t,a),l("body").on("mousemove touchmove",t,o),l("body").on("click mouseup mouseout touchend touchcancel",t,s),l(e).each(function(t,e){var url=e.getAttribute("src");l(e).parent().css("background-image","url("+url+")")})},window.ImageZoom=n},6205:function(t,e,i){"use strict";var n=i(6),TabsControl=i(137);window._npTabsInit=function(){function t(t){t.preventDefault(),t.stopPropagation();var link=n(t.currentTarget);new TabsControl(link).show()}n("body").on("click",".u-tab-link",t)},n(function(){window._npTabsInit()})},6206:function(t,e,i){"use strict";var n=i(6207);window._npLazyImages={setup:function(){window.lazySizesConfig=window.lazySizesConfig||{},window.lazySizesConfig.init=false,document.addEventListener("lazybeforeunveil",function(t){var el=t.target;if(el.matches("video")){var e=el.getAttribute("data-src"),i=el.querySelector("source");if(i&&e)i.setAttribute("src",e)}else{var n=el.getAttribute("data-bg");if(n){var list=cssBgParser.parseElementStyle(getComputedStyle(el));if(list.backgrounds.length)list.backgrounds[0].color="";list.backgrounds.push(new cssBgParser.Background({image:n})),el.style.backgroundImage=list.toString("image")}}})},init:function(){n.init()}},window._npLazyImages.setup(),$(function(){window._npLazyImages.init()})},6207:function(t,e,i){"use strict";!function(e,factory){var i=factory(e,e.document,Date);if(e.lazySizes=i,"object"==typeof t&&t.exports)t.exports=i}("undefined"!=typeof window?window:{},function t(e,i,n){var o,a;if(!function(){var t,i={lazyClass:"lazyload",loadedClass:"lazyloaded",loadingClass:"lazyloading",preloadClass:"lazypreload",errorClass:"lazyerror",autosizesClass:"lazyautosizes",srcAttr:"data-src",srcsetAttr:"data-srcset",sizesAttr:"data-sizes",minSize:40,customMedia:{},init:true,expFactor:1.5,hFac:.8,loadMode:2,loadHidden:true,ricTimeout:0,throttleDelay:125};a=e.lazySizesConfig||e.lazysizesConfig||{};for(t in i)if(!(t in a))a[t]=i[t]}(),!i||!i.getElementsByClassName)return{init:function(){},cfg:a,noSupport:true};var s=i.documentElement,l=e.HTMLPictureElement,u="addEventListener",f="getAttribute",c=e[u].bind(e),p=e.setTimeout,h=e.requestAnimationFrame||p,m=e.requestIdleCallback,v=/^picture$/i,g=["load","error","lazyincluded","_lazyloaded"],y={},w=Array.prototype.forEach,b=function(t,e){if(!y[e])y[e]=new RegExp("(\\s|^)"+e+"(\\s|$)");return y[e].test(t[f]("class")||"")&&y[e]},_=function(t,e){if(!b(t,e))t.setAttribute("class",(t[f]("class")||"").trim()+" "+e)},x=function(t,e){var i;if(i=b(t,e))t.setAttribute("class",(t[f]("class")||"").replace(i," "))},C=function(t,e,add){var i=add?u:"removeEventListener";if(add)C(t,e);g.forEach(function(n){t[i](n,e)})},k=function(t,e,n,a,s){var l=i.createEvent("Event");if(!n)n={};return n.instance=o,l.initEvent(e,!a,!s),l.detail=n,t.dispatchEvent(l),l},T=function(el,t){var i;if(!l&&(i=e.picturefill||a.pf)){if(t&&t.src&&!el[f]("srcset"))el.setAttribute("srcset",t.src);i({reevaluate:true,elements:[el]})}else if(t&&t.src)el.src=t.src},S=function(t,style){return(getComputedStyle(t,null)||{})[style]},A=function(t,e,i){for(i=i||t.offsetWidth;i<a.minSize&&e&&!t._lazysizesWidth;)i=e.offsetWidth,e=e.parentNode;return i},I=function(){var t,e,n=[],o=[],a=n,s=function(){var i=a;for(a=n.length?o:n,t=true,e=false;i.length;)i.shift()();t=false},l=function(n,o){if(t&&!o)n.apply(this,arguments);else if(a.push(n),!e)e=true,(i.hidden?p:h)(s)};return l._lsFlush=s,l}(),E=function(t,simple){return simple?function(){I(t)}:function(){var e=this,i=arguments;I(function(){t.apply(e,i)})}},L=function(t){var e,i=0,o=a.throttleDelay,s=a.ricTimeout,l=function(){e=false,i=n.now(),t()},u=m&&s>49?function(){if(m(l,{timeout:s}),s!==a.ricTimeout)s=a.ricTimeout}:E(function(){p(l)},true);return function(t){var a;if(t=true===t)s=33;if(!e){if(e=true,a=o-(n.now()-i),a<0)a=0;if(t||a<9)u();else p(u,a)}}},O=function(t){var e,i,o=99,a=function(){e=null,t()},s=function(){var t=n.now()-i;if(t<o)p(s,o-t);else(m||a)(a)};return function(){if(i=n.now(),!e)e=p(s,o)}},loader=function(){var t,l,m,g,y,A,M,z,P,N,H,B,W=/^img$/i,U=/^iframe$/i,V="onscroll"in e&&!/(gle|ing)bot/.test(navigator.userAgent),Z=0,j=0,X=0,$=-1,K=function(t){if(X--,!t||X<0||!t.target)X=0},G=function(t){if(null==B)B="hidden"==S(i.body,"visibility");return B||!("hidden"==S(t.parentNode,"visibility")&&"hidden"==S(t,"visibility"))},Y=function(t,e){var n,o=t,visible=G(t);for(z-=e,H+=e,P-=e,N+=e;visible&&(o=o.offsetParent)&&o!=i.body&&o!=s;)if(visible=(S(o,"opacity")||1)>0,visible&&"visible"!=S(o,"overflow"))n=o.getBoundingClientRect(),visible=N>n.left&&P<n.right&&H>n.top-1&&z<n.bottom+1;return visible},J=function(){var e,n,rect,u,c,p,h,m,v,y,w,b,_=o.elements;if((g=a.loadMode)&&X<8&&(e=_.length)){for(n=0,$++;n<e;n++)if(_[n]&&!_[n]._lazyRace)if(!(!V||o.prematureUnveil&&o.prematureUnveil(_[n]))){if(!(m=_[n][f]("data-expand"))||!(p=1*m))p=j;if(!y)if(y=!a.expand||a.expand<1?s.clientHeight>500&&s.clientWidth>500?500:370:a.expand,o._defEx=y,w=y*a.expFactor,b=a.hFac,B=null,j<w&&X<1&&$>2&&g>2&&!i.hidden)j=w,$=0;else if(g>1&&$>1&&X<6)j=y;else j=Z;if(v!==p)A=innerWidth+p*b,M=innerHeight+p,h=-1*p,v=p;if(rect=_[n].getBoundingClientRect(),(H=rect.bottom)>=h&&(z=rect.top)<=M&&(N=rect.right)>=h*b&&(P=rect.left)<=A&&(H||N||P||z)&&(a.loadHidden||G(_[n]))&&(l&&X<3&&!m&&(g<3||$<4)||Y(_[n],p))){if(ut(_[n]),c=true,X>9)break}else if(!c&&l&&!u&&X<4&&$<4&&g>2&&(t[0]||a.preloadAfterLoad)&&(t[0]||!m&&(H||N||P||z||"auto"!=_[n][f](a.sizesAttr))))u=t[0]||_[n]}else ut(_[n]);if(u&&!c)ut(u)}},tt=L(J),nt=function(t){var e=t.target;if(e._lazyCache)return delete e._lazyCache,void 0;K(t),_(e,a.loadedClass),x(e,a.loadingClass),C(e,rt),k(e,"lazyloaded")},ot=E(nt),rt=function(t){ot({target:t.target})},at=function(t,e){try{t.contentWindow.location.replace(e)}catch(i){t.src=e}},st=function(t){var e,i=t[f](a.srcsetAttr);if(e=a.customMedia[t[f]("data-media")||t[f]("media")])t.setAttribute("media",e);if(i)t.setAttribute("srcset",i)},lt=E(function(t,e,i,n,o){var s,l,u,c,h,g;if(!(h=k(t,"lazybeforeunveil",e)).defaultPrevented){if(n)if(i)_(t,a.autosizesClass);else t.setAttribute("sizes",n);if(l=t[f](a.srcsetAttr),s=t[f](a.srcAttr),o)u=t.parentNode,c=u&&v.test(u.nodeName||"");if(g=e.firesLoad||"src"in t&&(l||s||c),h={target:t},_(t,a.loadingClass),g)clearTimeout(m),m=p(K,2500),C(t,rt,true);if(c)w.call(u.getElementsByTagName("source"),st);if(l)t.setAttribute("srcset",l);else if(s&&!c)if(U.test(t.nodeName))at(t,s);else t.src=s;if(o&&(l||c))T(t,{src:s})}if(t._lazyRace)delete t._lazyRace;x(t,a.lazyClass),I(function(){var e=t.complete&&t.naturalWidth>1;if(!g||e){if(e)_(t,"ls-is-cached");nt(h),t._lazyCache=true,p(function(){if("_lazyCache"in t)delete t._lazyCache},9)}if("lazy"==t.loading)X--},true)}),ut=function(t){if(!t._lazyRace){var e,i=W.test(t.nodeName),n=i&&(t[f](a.sizesAttr)||t[f]("sizes")),o="auto"==n;if(!o&&l||!i||!t[f]("src")&&!t.srcset||t.complete||b(t,a.errorClass)||!b(t,a.lazyClass)){if(e=k(t,"lazyunveilread").detail,o)F.updateElem(t,true,t.offsetWidth);t._lazyRace=true,X++,lt(t,e,o,n,i)}}},ft=O(function(){a.loadMode=3,tt()}),ct=function(){if(3==a.loadMode)a.loadMode=2;ft()},dt=function(){if(!l){if(n.now()-y<999)return p(dt,999),void 0;l=true,a.loadMode=3,tt(),c("scroll",ct,true)}};return{_:function(){if(y=n.now(),o.elements=i.getElementsByClassName(a.lazyClass),t=i.getElementsByClassName(a.lazyClass+" "+a.preloadClass),c("scroll",tt,true),c("resize",tt,true),c("pageshow",function(t){if(t.persisted){var e=i.querySelectorAll("."+a.loadingClass);if(e.length&&e.forEach)h(function(){e.forEach(function(t){if(t.complete)ut(t)})})}}),e.MutationObserver)new MutationObserver(tt).observe(s,{childList:true,subtree:true,attributes:true});else s[u]("DOMNodeInserted",tt,true),s[u]("DOMAttrModified",tt,true),setInterval(tt,999);if(c("hashchange",tt,true),["focus","mouseover","click","load","transitionend","animationend"].forEach(function(t){i[u](t,tt,true)}),/d$|^c/.test(i.readyState))dt();else c("load",dt),i[u]("DOMContentLoaded",tt),p(dt,2e4);if(o.elements.length)J(),I._lsFlush();else tt()},checkElems:tt,unveil:ut,_aLSL:ct}}(),F=function(){var t,e=E(function(t,e,i,n){var o,a,s;if(t._lazysizesWidth=n,n+="px",t.setAttribute("sizes",n),v.test(e.nodeName||""))for(o=e.getElementsByTagName("source"),a=0,s=o.length;a<s;a++)o[a].setAttribute("sizes",n);if(!i.detail.dataAttr)T(t,i.detail)}),n=function(t,i,n){var o,a=t.parentNode;if(a)if(n=A(t,a,n),o=k(t,"lazybeforesizes",{width:n,dataAttr:!!i}),!o.defaultPrevented)if(n=o.detail.width,n&&n!==t._lazysizesWidth)e(t,a,o,n)},o=function(){var e,i=t.length;if(i)for(e=0;e<i;e++)n(t[e])},s=O(o);return{_:function(){t=i.getElementsByClassName(a.autosizesClass),c("resize",s)},checkElems:s,updateElem:n}}(),init=function(){if(!init.i&&i.getElementsByClassName)init.i=true,F._(),loader._()};return p(function(){if(a.init)init()}),o={cfg:a,autoSizer:F,loader:loader,init:init,uP:T,aC:_,rC:x,hC:b,fire:k,gW:A,rAF:I},o})},6208:function(t,e,i){"use strict";var n=i(6),Dialog=i(101);window._npDialogsInit=function(){function t(t){t.preventDefault(),t.stopPropagation(),i(t).open()}function e(t){t.preventDefault(),t.stopPropagation(),i(t).close()}function i(t){var link=n(t.currentTarget),e=link.attr("href")||link.attr("data-href"),i=n(e);return i=i.length?i:link,new Dialog(i)}function o(){return new Dialog(n('[data-dialog-show-on="page_exit"]'))}function a(){return new Dialog(n('[data-dialog-show-on="timer"]'))}function s(t){if(t.clientY<50&&null==t.relatedTarget&&"select"!==t.target.nodeName.toLowerCase()){o().open(function(){document.removeEventListener("mouseout",s)})}}function l(){var dialog=a();setTimeout(function(){dialog.open()},dialog.getInterval())}function u(t){var e=n(t.currentTarget);setTimeout(function(){new Dialog(e).close()})}n("body").on("click",".u-dialog-link",t),n("body").on("click",".u-dialog-close-button",e),n("body").on("click",".u-dialog .u-btn",u),document.addEventListener("mouseout",s),l()},n(function(){window._npDialogsInit()})},6209:function(t,e,i){"use strict";var n=i(6);n(function(){n(document).on("click",".u-quantity-input a",function(t){t.preventDefault();var e,i=n(this),o=i.siblings("input");if(i.hasClass("minus"))e=parseFloat(o.val())-1,e=e<1?1:e,o.val(e);if(i.hasClass("plus"))e=parseFloat(o.val())+1,o.val(e);i.siblings(".minus").addBack(".minus").toggleClass("disabled",1===e),o.change()})})},6210:function(t,e,i){"use strict";var n=i(6),Accordion=i(81);window._npAccordionInit=function(){function t(t){t.preventDefault(),t.stopPropagation();var link=n(t.currentTarget);new Accordion(link).show()}n("body").on("click",".u-accordion-link",t)},n(function(){window._npAccordionInit()})},6211:function(t,e){},81:function(t,e,i){"use strict";function Accordion(link){this.selector=".u-accordion",this.activeClass="u-accordion-active",this._paneSelector=".u-accordion-pane",this.activeSelector="."+this.activeClass,this._linkSelector=".u-accordion-link",this.activeLinkClass="active",this.activeLinkSelector="."+this.activeLinkClass,this._isCollapsedByDefaultSelector=".u-collapsed-by-default",this._link=link,this._accordion=this._link.closest(this.selector)}t.exports=Accordion,Accordion.prototype.show=function(t){var link=this._link;if(link.is(this.activeLinkSelector)&&!t)return this._removeActiveLink(),this._hidePane(link),void 0;this._removeActiveLink(),this._hidePane(link),this._addActiveLink(link),this._activatePane(link)},Accordion.prototype._removeActiveLink=function(){var t=this._getActiveLink();t.removeClass(this.activeLinkClass),t.attr("aria-selected",false)},Accordion.prototype._getActiveLink=function(){return this._accordion.find(this.activeLinkSelector)},Accordion.prototype._addActiveLink=function(link){link.addClass(this.activeLinkClass),link.attr("aria-selected",true)},Accordion.prototype._activatePane=function(link){this._accordion.find(this.activeSelector).removeClass(this.activeClass),this._getPane(link).addClass(this.activeClass)},Accordion.prototype._getPane=function(link){return link.siblings(this._paneSelector)},Accordion.prototype._hidePane=function(link){this._getPane(link).removeClass(this.activeClass)},Accordion.prototype.closeAll=function(){this._accordion.find(this._linkSelector+this.activeLinkSelector).removeClass(this.activeLinkClass).attr("aria-selected",false),this._accordion.find(this._paneSelector+this.activeSelector).removeClass(this.activeClass)},Accordion.prototype.isCollapsedByDefault=function(){return this._accordion.is(this._isCollapsedByDefaultSelector)}}});
-window.onerror = function(msg) {
-    alert(msg + ' See console (F12) for details. \nReport issues to https://github.com/watchyoursixx/HunterSim/issues');
-    return true;
-}
+//window.onerror = function(msg) {
+//    alert(msg + ' See console (F12) for details. \nReport issues to https://github.com/watchyoursixx/HunterSim/issues');
+//    return true;
+//}
