+++ react/components/DialogTable/useStyles.js
@@ -1,0 +1,16 @@
+import { makeStyles } from "@mui/styles";
+
+export const useStyles = makeStyles((theme) => ({
+  Epic: {
+    color: "#a335ee",
+  },
+  Rare: {
+    color: "#0070dd",
+  },
+  Uncommon: {
+    color: "#1eff00",
+  },
+  Common: {
+    color: "#000000",
+  }
+}));
