+++ react/components/Wowhead/useStyles.js
@@ -1,0 +1,8 @@
+import { makeStyles } from "@mui/styles";
+
+export const useStyles = makeStyles((theme) => ({
+img : {
+    border : '1px solid #5f5f63'
+}
+}));
+
