+++ react/constants/gearKeys.js
@@ -1,0 +1,20 @@
+const GEAR_KEYS = {
+    MELEE_WEAPONS : 'MELEE_WEAPONS',
+    AMMOS : 'AMMOS',
+    BACKS: "BACKS",
+    CHESTS: "CHESTS",
+    GEAR_MAP: "GEAR_MAP",
+    FEET: "FEET",
+    HANDS: "HANDS",
+    HEADS: "HEADS",
+    QUIVERS: "QUIVERS",
+    LEGS: "LEGS",
+    SHOULDERS: "SHOULDERS",
+    TRINKETS: "TRINKETS",
+    WAISTS: "WAISTS",
+    RANGED_WEAPONS: "RANGED_WEAPONS",
+    WRISTS: "WRISTS",
+    RINGS: "RINGS",
+    NECKS: "NECKS",
+    SETS: "SETS",
+}
